channel

+++ db/src/user.rs
@@ -10,7 +10,7 @@ impl UserQuery {
         User::find().all(db).await
     }
 
-    pub async fn find_by_id(db: &DbConn, id: i32) -> Result<Option<user::Model>, DbErr> {
+    pub async fn find_by_id(db: &DbConn, id: i64) -> Result<Option<user::Model>, DbErr> {
         User::find_by_id(id).one(db).await
     }
